requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
pyreadline3==3.5.4
pytest==8.0.0
requests==2.33.0
urllib3==2.6.3